Add token duplication check to TwitterData objects
Added methods to read and write data token ids to their respective CSV files; methods are general and are inherited in Tweets, Users, and Places. The 'id' column is checked for entries, meaning metadata located in 'twitterdata' files will throw an exception and it's not currently handled!

Added support for 'pytest' testing

Restructured file organization to adhere to design standards